Moved layers classes into a layer folder
